fix(codacy): isolate reserve scenario persistence helpers
Codacy's second pass was limited to changed-file size and method metric attribution. Move snapshot read/write helpers into a focused store module and keep the reserve summary allocation cap helper outside the long summary builder region without changing runtime behavior.

Constraint: Codacy flags changed-file size and parser-attributed method metrics on PR #714.
Rejected: Suppress remaining TypeScript complexity issues | structural extraction keeps the production contract explicit.
Confidence: high
Scope-risk: narrow
Reversibility: clean
Tested: npx vitest run tests/unit/services/fund-scenario-reserve-summary.test.ts tests/unit/services/fund-scenario-reserve-calculation-service.test.ts tests/unit/workers/fund-scenario-calc-worker.test.ts --project=server
Tested: npm run check
Tested: npm run lint
Tested: npm run test:unit
Tested: git diff --check
Not-tested: Live BullMQ/Redis worker execution.
